brain/requirements.txt
jupyterlab
jsonlines
# For local development
doppler-env
# Google Cloud Services
google-cloud-bigquery
google-cloud-storage